+++ packages/diagram/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inkset/diagram
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @inkset/core@0.1.4
+
 ## 0.1.4
 
 ### Patch Changes