+++ packages/math/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inkset/math
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @inkset/core@0.1.4
+
 ## 0.1.3
 
 ### Patch Changes